move ChannelWithVideos to new directory
